LPS-107005 Move reindex logic to kaleo integration module
