+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-metrics-integration/src/main/java/com/liferay/portal/workflow/kaleo/metrics/integration/internal/search/index/reindexer/InstanceWorkflowMetricsReindexer.java
@@ -1,0 +1,102 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.portal.workflow.kaleo.metrics.integration.internal.search.index.reindexer;
+
+import com.liferay.asset.kernel.service.AssetEntryLocalService;
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.workflow.kaleo.metrics.integration.internal.util.InstanceIndexerHelper;
+import com.liferay.portal.workflow.kaleo.model.KaleoDefinitionVersion;
+import com.liferay.portal.workflow.kaleo.model.KaleoInstance;
+import com.liferay.portal.workflow.kaleo.service.KaleoDefinitionVersionLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoInstanceLocalService;
+import com.liferay.portal.workflow.metrics.search.index.InstanceWorkflowMetricsIndexer;
+import com.liferay.portal.workflow.metrics.search.index.reindexer.WorkflowMetricsReindexer;
+
+import java.util.Objects;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Rafael Praxedes
+ */
+@Component(
+	immediate = true, property = "workflow.metrics.index.entity.name=instance",
+	service = WorkflowMetricsReindexer.class
+)
+public class InstanceWorkflowMetricsReindexer
+	implements WorkflowMetricsReindexer {
+
+	@Override
+	public void reindex(long companyId) throws PortalException {
+		InstanceIndexerHelper instanceIndexerHelper = new InstanceIndexerHelper(
+			_assetEntryLocalService);
+
+		ActionableDynamicQuery actionableDynamicQuery =
+			_kaleoInstanceLocalService.getActionableDynamicQuery();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			dynamicQuery -> {
+				Property companyIdProperty = PropertyFactoryUtil.forName(
+					"companyId");
+
+				dynamicQuery.add(companyIdProperty.eq(companyId));
+			});
+		actionableDynamicQuery.setPerformActionMethod(
+			(KaleoInstance kaleoInstance) -> {
+				KaleoDefinitionVersion kaleoDefinitionVersion =
+					_kaleoDefinitionVersionLocalService.
+						fetchKaleoDefinitionVersion(
+							kaleoInstance.getKaleoDefinitionVersionId());
+
+				if (Objects.isNull(kaleoDefinitionVersion)) {
+					return;
+				}
+
+				_instanceWorkflowMetricsIndexer.addInstance(
+					instanceIndexerHelper.createAssetTitleLocalizationMap(
+						kaleoInstance),
+					instanceIndexerHelper.createAssetTypeLocalizationMap(
+						kaleoInstance),
+					kaleoInstance.getClassName(), kaleoInstance.getClassPK(),
+					companyId, kaleoInstance.getCompletionDate(),
+					kaleoInstance.getCreateDate(),
+					kaleoInstance.getKaleoInstanceId(),
+					kaleoInstance.getModifiedDate(),
+					kaleoInstance.getKaleoDefinitionId(),
+					kaleoDefinitionVersion.getVersion(),
+					kaleoInstance.getUserId(), kaleoInstance.getUserName());
+			});
+
+		actionableDynamicQuery.performActions();
+	}
+
+	@Reference
+	private AssetEntryLocalService _assetEntryLocalService;
+
+	@Reference
+	private InstanceWorkflowMetricsIndexer _instanceWorkflowMetricsIndexer;
+
+	@Reference
+	private KaleoDefinitionVersionLocalService
+		_kaleoDefinitionVersionLocalService;
+
+	@Reference
+	private KaleoInstanceLocalService _kaleoInstanceLocalService;
+
+}
