+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-metrics-integration/src/main/java/com/liferay/portal/workflow/kaleo/metrics/integration/internal/search/index/reindexer/NodeWorkflowMetricsReindexer.java
@@ -1,0 +1,140 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.portal.workflow.kaleo.metrics.integration.internal.search.index.reindexer;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.workflow.kaleo.definition.NodeType;
+import com.liferay.portal.workflow.kaleo.model.KaleoDefinitionVersion;
+import com.liferay.portal.workflow.kaleo.model.KaleoNode;
+import com.liferay.portal.workflow.kaleo.model.KaleoTask;
+import com.liferay.portal.workflow.kaleo.service.KaleoDefinitionVersionLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoNodeLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoTaskLocalService;
+import com.liferay.portal.workflow.metrics.search.index.NodeWorkflowMetricsIndexer;
+import com.liferay.portal.workflow.metrics.search.index.reindexer.WorkflowMetricsReindexer;
+
+import java.util.Objects;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Rafael Praxedes
+ */
+@Component(
+	immediate = true, property = "workflow.metrics.index.entity.name=node",
+	service = WorkflowMetricsReindexer.class
+)
+public class NodeWorkflowMetricsReindexer implements WorkflowMetricsReindexer {
+
+	@Override
+	public void reindex(long companyId) throws PortalException {
+		_reindexIndexWithKaleoNode(companyId);
+		_reindexIndexWithKaleoTask(companyId);
+	}
+
+	private void _reindexIndexWithKaleoNode(long companyId)
+		throws PortalException {
+
+		ActionableDynamicQuery actionableDynamicQuery =
+			_kaleoNodeLocalService.getActionableDynamicQuery();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			dynamicQuery -> {
+				Property companyIdProperty = PropertyFactoryUtil.forName(
+					"companyId");
+
+				dynamicQuery.add(companyIdProperty.eq(companyId));
+
+				Property typeProperty = PropertyFactoryUtil.forName("type");
+
+				dynamicQuery.add(typeProperty.eq(NodeType.STATE.name()));
+			});
+		actionableDynamicQuery.setPerformActionMethod(
+			(KaleoNode kaleoNode) -> {
+				KaleoDefinitionVersion kaleoDefinitionVersion =
+					_kaleoDefinitionVersionLocalService.
+						fetchKaleoDefinitionVersion(
+							kaleoNode.getKaleoDefinitionVersionId());
+
+				if (Objects.isNull(kaleoDefinitionVersion)) {
+					return;
+				}
+
+				_nodeWorkflowMetricsIndexer.addNode(
+					kaleoNode.getCompanyId(), kaleoNode.getCreateDate(),
+					kaleoNode.isInitial(), kaleoNode.getModifiedDate(),
+					kaleoNode.getName(), kaleoNode.getKaleoNodeId(),
+					kaleoNode.getKaleoDefinitionId(),
+					kaleoDefinitionVersion.getVersion(), kaleoNode.isTerminal(),
+					kaleoNode.getType());
+			});
+
+		actionableDynamicQuery.performActions();
+	}
+
+	private void _reindexIndexWithKaleoTask(long companyId)
+		throws PortalException {
+
+		ActionableDynamicQuery actionableDynamicQuery =
+			_kaleoTaskLocalService.getActionableDynamicQuery();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			dynamicQuery -> {
+				Property companyIdProperty = PropertyFactoryUtil.forName(
+					"companyId");
+
+				dynamicQuery.add(companyIdProperty.eq(companyId));
+			});
+		actionableDynamicQuery.setPerformActionMethod(
+			(KaleoTask kaleoTask) -> {
+				KaleoDefinitionVersion kaleoDefinitionVersion =
+					_kaleoDefinitionVersionLocalService.
+						fetchKaleoDefinitionVersion(
+							kaleoTask.getKaleoDefinitionVersionId());
+
+				if (Objects.isNull(kaleoDefinitionVersion)) {
+					return;
+				}
+
+				_nodeWorkflowMetricsIndexer.addNode(
+					kaleoTask.getCompanyId(), kaleoTask.getCreateDate(), false,
+					kaleoTask.getModifiedDate(), kaleoTask.getName(),
+					kaleoTask.getKaleoTaskId(),
+					kaleoTask.getKaleoDefinitionId(),
+					kaleoDefinitionVersion.getVersion(), false,
+					NodeType.TASK.name());
+			});
+
+		actionableDynamicQuery.performActions();
+	}
+
+	@Reference
+	private KaleoDefinitionVersionLocalService
+		_kaleoDefinitionVersionLocalService;
+
+	@Reference
+	private KaleoNodeLocalService _kaleoNodeLocalService;
+
+	@Reference
+	private KaleoTaskLocalService _kaleoTaskLocalService;
+
+	@Reference
+	private NodeWorkflowMetricsIndexer _nodeWorkflowMetricsIndexer;
+
+}
