+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-metrics-integration/src/main/java/com/liferay/portal/workflow/kaleo/metrics/integration/internal/search/index/reindexer/TaskWorkflowMetricsReindexer.java
@@ -1,0 +1,133 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.portal.workflow.kaleo.metrics.integration.internal.search.index.reindexer;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.workflow.kaleo.model.KaleoDefinitionVersion;
+import com.liferay.portal.workflow.kaleo.model.KaleoInstance;
+import com.liferay.portal.workflow.kaleo.model.KaleoTaskAssignmentInstance;
+import com.liferay.portal.workflow.kaleo.model.KaleoTaskInstanceToken;
+import com.liferay.portal.workflow.kaleo.service.KaleoDefinitionVersionLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoInstanceLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoTaskAssignmentInstanceLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoTaskInstanceTokenLocalService;
+import com.liferay.portal.workflow.metrics.search.index.TaskWorkflowMetricsIndexer;
+import com.liferay.portal.workflow.metrics.search.index.reindexer.WorkflowMetricsReindexer;
+
+import java.util.Objects;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Rafael Praxedes
+ */
+@Component(
+	immediate = true, property = "workflow.metrics.index.entity.name=task",
+	service = WorkflowMetricsReindexer.class
+)
+public class TaskWorkflowMetricsReindexer implements WorkflowMetricsReindexer {
+
+	@Override
+	public void reindex(long companyId) throws PortalException {
+		ActionableDynamicQuery actionableDynamicQuery =
+			_kaleoTaskInstanceTokenLocalService.getActionableDynamicQuery();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			dynamicQuery -> {
+				Property companyIdProperty = PropertyFactoryUtil.forName(
+					"companyId");
+
+				dynamicQuery.add(companyIdProperty.eq(companyId));
+			});
+		actionableDynamicQuery.setPerformActionMethod(
+			(KaleoTaskInstanceToken kaleoTaskInstanceToken) -> {
+				KaleoDefinitionVersion kaleoDefinitionVersion =
+					_kaleoDefinitionVersionLocalService.
+						fetchKaleoDefinitionVersion(
+							kaleoTaskInstanceToken.
+								getKaleoDefinitionVersionId());
+
+				if (Objects.isNull(kaleoDefinitionVersion)) {
+					return;
+				}
+
+				KaleoInstance kaleoInstance =
+					_kaleoInstanceLocalService.fetchKaleoInstance(
+						kaleoTaskInstanceToken.getKaleoInstanceId());
+
+				if (Objects.isNull(kaleoInstance)) {
+					return;
+				}
+
+				KaleoTaskAssignmentInstance kaleoTaskAssignmentInstance =
+					_kaleoTaskAssignmentInstanceLocalService.
+						fetchFirstKaleoTaskAssignmentInstance(
+							kaleoTaskInstanceToken.
+								getKaleoTaskInstanceTokenId(),
+							User.class.getName(), null);
+
+				Long assigneeId = null;
+
+				if (kaleoTaskAssignmentInstance != null) {
+					assigneeId =
+						kaleoTaskAssignmentInstance.getAssigneeClassPK();
+				}
+
+				_taskWorkflowMetricsIndexer.addTask(
+					assigneeId, kaleoTaskInstanceToken.getClassName(),
+					kaleoTaskInstanceToken.getClassPK(),
+					kaleoTaskInstanceToken.getCompanyId(),
+					kaleoTaskInstanceToken.isCompleted(),
+					kaleoTaskInstanceToken.getCompletionDate(),
+					kaleoTaskInstanceToken.getCompletionUserId(),
+					kaleoTaskInstanceToken.getCreateDate(),
+					kaleoInstance.isCompleted(),
+					kaleoTaskInstanceToken.getKaleoInstanceId(),
+					kaleoTaskInstanceToken.getModifiedDate(),
+					kaleoTaskInstanceToken.getKaleoTaskName(),
+					kaleoTaskInstanceToken.getKaleoTaskId(),
+					kaleoInstance.getKaleoDefinitionId(),
+					kaleoDefinitionVersion.getVersion(),
+					kaleoTaskInstanceToken.getKaleoTaskInstanceTokenId(),
+					kaleoTaskInstanceToken.getUserId());
+			});
+
+		actionableDynamicQuery.performActions();
+	}
+
+	@Reference
+	private KaleoDefinitionVersionLocalService
+		_kaleoDefinitionVersionLocalService;
+
+	@Reference
+	private KaleoInstanceLocalService _kaleoInstanceLocalService;
+
+	@Reference
+	private KaleoTaskAssignmentInstanceLocalService
+		_kaleoTaskAssignmentInstanceLocalService;
+
+	@Reference
+	private KaleoTaskInstanceTokenLocalService
+		_kaleoTaskInstanceTokenLocalService;
+
+	@Reference
+	private TaskWorkflowMetricsIndexer _taskWorkflowMetricsIndexer;
+
+}
