+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-metrics-integration/src/main/java/com/liferay/portal/workflow/kaleo/metrics/integration/internal/search/index/reindexer/TransitionWorkflowMetricsReindexer.java
@@ -1,0 +1,123 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.portal.workflow.kaleo.metrics.integration.internal.search.index.reindexer;
+
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.workflow.kaleo.definition.NodeType;
+import com.liferay.portal.workflow.kaleo.model.KaleoDefinitionVersion;
+import com.liferay.portal.workflow.kaleo.model.KaleoNode;
+import com.liferay.portal.workflow.kaleo.model.KaleoTask;
+import com.liferay.portal.workflow.kaleo.model.KaleoTransition;
+import com.liferay.portal.workflow.kaleo.service.KaleoDefinitionVersionLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoNodeLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoTaskLocalService;
+import com.liferay.portal.workflow.kaleo.service.KaleoTransitionLocalService;
+import com.liferay.portal.workflow.metrics.search.index.TransitionWorkflowMetricsIndexer;
+import com.liferay.portal.workflow.metrics.search.index.reindexer.WorkflowMetricsReindexer;
+
+import java.util.Objects;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Rafael Praxedes
+ */
+@Component(
+	immediate = true,
+	property = "workflow.metrics.index.entity.name=transition",
+	service = WorkflowMetricsReindexer.class
+)
+public class TransitionWorkflowMetricsReindexer
+	implements WorkflowMetricsReindexer {
+
+	@Override
+	public void reindex(long companyId) throws PortalException {
+		ActionableDynamicQuery actionableDynamicQuery =
+			_kaleoTransitionLocalService.getActionableDynamicQuery();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			dynamicQuery -> {
+				Property companyIdProperty = PropertyFactoryUtil.forName(
+					"companyId");
+
+				dynamicQuery.add(companyIdProperty.eq(companyId));
+			});
+		actionableDynamicQuery.setPerformActionMethod(
+			(KaleoTransition kaleoTransition) -> {
+				KaleoDefinitionVersion kaleoDefinitionVersion =
+					_kaleoDefinitionVersionLocalService.
+						fetchKaleoDefinitionVersion(
+							kaleoTransition.getKaleoDefinitionVersionId());
+
+				if (Objects.isNull(kaleoTransition)) {
+					return;
+				}
+
+				_transitionWorkflowMetricsIndexer.addTransition(
+					kaleoTransition.getCompanyId(),
+					kaleoTransition.getCreateDate(),
+					kaleoTransition.getModifiedDate(),
+					kaleoTransition.getName(),
+					_getNodeId(kaleoTransition.getKaleoNodeId()),
+					kaleoTransition.getKaleoDefinitionId(),
+					kaleoDefinitionVersion.getVersion(),
+					_getNodeId(kaleoTransition.getSourceKaleoNodeId()),
+					kaleoTransition.getSourceKaleoNodeName(),
+					_getNodeId(kaleoTransition.getTargetKaleoNodeId()),
+					kaleoTransition.getTargetKaleoNodeName(),
+					kaleoTransition.getKaleoTransitionId(),
+					kaleoTransition.getUserId());
+			});
+
+		actionableDynamicQuery.performActions();
+	}
+
+	private long _getNodeId(long kaleoNodeId) throws PortalException {
+		KaleoNode kaleoNode = _kaleoNodeLocalService.fetchKaleoNode(
+			kaleoNodeId);
+
+		if ((kaleoNode == null) ||
+			!Objects.equals(kaleoNode.getType(), NodeType.TASK.name())) {
+
+			return kaleoNodeId;
+		}
+
+		KaleoTask kaleoTask = _kaleoTaskLocalService.getKaleoNodeKaleoTask(
+			kaleoNode.getKaleoNodeId());
+
+		return kaleoTask.getKaleoTaskId();
+	}
+
+	@Reference
+	private KaleoDefinitionVersionLocalService
+		_kaleoDefinitionVersionLocalService;
+
+	@Reference
+	private KaleoNodeLocalService _kaleoNodeLocalService;
+
+	@Reference
+	private KaleoTaskLocalService _kaleoTaskLocalService;
+
+	@Reference
+	private KaleoTransitionLocalService _kaleoTransitionLocalService;
+
+	@Reference
+	private TransitionWorkflowMetricsIndexer _transitionWorkflowMetricsIndexer;
+
+}
