+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-metrics-integration/src/main/java/com/liferay/portal/workflow/kaleo/metrics/integration/internal/util/InstanceIndexerHelper.java
@@ -1,0 +1,127 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * The contents of this file are subject to the terms of the Liferay Enterprise
+ * Subscription License ("License"). You may not use this file except in
+ * compliance with the License. You can obtain a copy of the License by
+ * contacting Liferay, Inc. See the License for the specific language governing
+ * permissions and limitations under the License, including but not limited to
+ * distribution rights of the Software.
+ *
+ *
+ *
+ */
+
+package com.liferay.portal.workflow.kaleo.metrics.integration.internal.util;
+
+import com.liferay.asset.kernel.AssetRendererFactoryRegistryUtil;
+import com.liferay.asset.kernel.model.AssetEntry;
+import com.liferay.asset.kernel.model.AssetRenderer;
+import com.liferay.asset.kernel.model.AssetRendererFactory;
+import com.liferay.asset.kernel.service.AssetEntryLocalService;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.security.permission.ResourceActionsUtil;
+import com.liferay.portal.kernel.util.LocalizationUtil;
+import com.liferay.portal.kernel.workflow.WorkflowHandler;
+import com.liferay.portal.kernel.workflow.WorkflowHandlerRegistryUtil;
+import com.liferay.portal.workflow.kaleo.model.KaleoInstance;
+
+import java.util.HashMap;
+import java.util.Locale;
+import java.util.Map;
+
+/**
+ * @author Rafael Praxedes
+ */
+public class InstanceIndexerHelper {
+
+	public InstanceIndexerHelper(
+		AssetEntryLocalService assetEntryLocalService) {
+
+		_assetEntryLocalService = assetEntryLocalService;
+	}
+
+	public Map<Locale, String> createAssetTitleLocalizationMap(
+		KaleoInstance kaleoInstance) {
+
+		try {
+			AssetRenderer<?> assetRenderer = _getAssetRenderer(
+				kaleoInstance.getClassName(), kaleoInstance.getClassPK());
+
+			if (assetRenderer != null) {
+				AssetEntry assetEntry = _assetEntryLocalService.getEntry(
+					assetRenderer.getClassName(), assetRenderer.getClassPK());
+
+				return LocalizationUtil.populateLocalizationMap(
+					assetEntry.getTitleMap(), assetEntry.getDefaultLanguageId(),
+					assetEntry.getGroupId());
+			}
+		}
+		catch (PortalException portalException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(portalException, portalException);
+			}
+		}
+
+		WorkflowHandler<?> workflowHandler =
+			WorkflowHandlerRegistryUtil.getWorkflowHandler(
+				kaleoInstance.getClassName());
+
+		Map<Locale, String> localizationMap = new HashMap<>();
+
+		for (Locale availableLocale :
+				LanguageUtil.getAvailableLocales(kaleoInstance.getGroupId())) {
+
+			localizationMap.put(
+				availableLocale,
+				workflowHandler.getTitle(
+					kaleoInstance.getClassPK(), availableLocale));
+		}
+
+		return localizationMap;
+	}
+
+	public Map<Locale, String> createAssetTypeLocalizationMap(
+		KaleoInstance kaleoInstance) {
+
+		Map<Locale, String> localizationMap = new HashMap<>();
+
+		for (Locale availableLocale :
+				LanguageUtil.getAvailableLocales(kaleoInstance.getGroupId())) {
+
+			localizationMap.put(
+				availableLocale,
+				ResourceActionsUtil.getModelResource(
+					availableLocale, kaleoInstance.getClassName()));
+		}
+
+		return localizationMap;
+	}
+
+	private AssetRenderer<?> _getAssetRenderer(String className, long classPK)
+		throws PortalException {
+
+		AssetRendererFactory<?> assetRendererFactory = _getAssetRendererFactory(
+			className);
+
+		if (assetRendererFactory != null) {
+			return assetRendererFactory.getAssetRenderer(classPK);
+		}
+
+		return null;
+	}
+
+	private AssetRendererFactory<?> _getAssetRendererFactory(String className) {
+		return AssetRendererFactoryRegistryUtil.
+			getAssetRendererFactoryByClassName(className);
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		InstanceIndexerHelper.class);
+
+	private final AssetEntryLocalService _assetEntryLocalService;
+
+}
